pyproject.toml
[project]
name = "codeshot"
version = "1.0.0"
description = "Beautiful code screenshot generator with themes, effects, and professional styling"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Codeshot Team", email = "ayyan.sameer.shaikh.eren@gmail.com"}
]
keywords = ["code", "screenshot", "image", "generator", "syntax-highlighting"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.8.0",
    "pytesseract>=0.3.10",
    "pygments>=2.18.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pytest>=8.4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
codeshot = "main:main"

[project.urls]
Homepage = "https://github.com/codeshot-team/codeshot"
Repository = "https://github.com/codeshot-team/codeshot"
Documentation = "https://docs.codeshot.dev"
"Bug Tracker" = "https://github.com/codeshot-team/codeshot/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config", "fonts"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
